requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
discord.py>=2.0.0
openai>=0.27.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
pytz>=2021.1
pytest>=6.0.0